// kitchen-planner-fe-main/src/app/api/api-kms-planner-masterdata/model/billingInfo.ts
/**
 * KMS - Kitchen Management System API
 *
 * Contact: anas-ur-rasheed.abid@tu-ilmenau.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Address } from './address';


export interface BillingInfo { 
    firstName?: string;
    lastName?: string;
    email?: string;
    phone?: string;
    address?: Address;
    company?: string;
    taxId?: string;
}