// kitchen-planner-fe-main/src/app/api/api-kms-planner-masterdata/model/userProfile.ts
/**
 * KMS - Kitchen Management System API
 *
 * Contact: anas-ur-rasheed.abid@tu-ilmenau.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserProfile { 
    _id?: string;
    firstName?: string;
    lastName?: string;
    email?: string;
    role?: string;
    createdAt?: string;
}